requirements.txt
alembic==1.7.6
anyio==3.6.1
attrs==21.4.0
autopep8==1.6.0
azure-ai-textanalytics==5.2.0b1
azure-common==1.1.28
azure-core==1.24.0
bcrypt==3.2.0
black==22.1.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
djlint==1.0.3
dnspython==1.16.0
eventlet==0.31.0
fastapi==0.79.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
amqp==5.1.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
greenlet==1.1.2
kombu==5.2.4
prompt-toolkit==3.0.30
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
vine==5.0.0
wcwidth==0.2.5
importlib-metadata==4.11.2
importlib-resources==5.4.0
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.1
Mako==1.1.6
MarkupSafe==2.1.0
msrest==0.6.21
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.19.2
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.18.2
Werkzeug==2.0.3
WTForms==3.0.1
zipp==3.7.0
anyio==3.6.1
asgiref==3.5.2
beautifulsoup4==4.11.1
blis==0.7.7
bs4==0.0.1
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
de-core-news-md==3.3.0
en-core-web-md==3.3.0
es-core-news-md==3.3.0
et-xmlfile==1.1.0
fastapi==0.78.0
fastapi-versioning==0.10.0
filelock==3.7.0
fr-core-news-md==3.3.0
h11==0.13.0
httptools==0.4.0
huggingface-hub==0.7.0
idna==3.3
it-core-news-md==3.3.0
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pandas==1.4.2
pathy==0.6.1
Pillow==9.1.1
preshed==3.0.6
pt-core-news-md==3.3.0
pydantic==1.8.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.0
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
soupsieve==2.3.2.post1
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.3
starlette==0.19.1
thinc==8.0.16
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
transformers==4.19.2
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
wasabi==0.9.1
watchgod==0.8.2
websockets==10.3
xmltodict==0.13.0